Refactoring addressing comments: AddMethod CreateAntlrLexer(), move PreprocessText() to AntlrLexer
