--- Sources/PT.PM.Common/ILanguageHandler.cs
@@ -1,7 +1,0 @@
-namespace PT.PM.Common
-{
-    public interface ILanguageHandler : ILoggable
-    {
-        Language Language { get; }
-    }
-}
